Bug 1147491, part 2 - Disable use of BaseTimeDuration::operator/ with floating point values. r=ehsan

+++ widget/gonk/GeckoTouchDispatcher.cpp
@@ -271,7 +271,7 @@ GoannaTouchDispatcher::ResampleTouchMoves(MultiTouchInput& aOutTouch, TimeStamp 
   TimeDuration touchDiff = currentTouch.mTimeStamp - baseTouch.mTimeStamp;
 
   if (currentTouch.mTimeStamp < sampleTime) {
-    TimeDuration maxResampleTime = std::min(touchDiff / 2, mMaxPredict);
+    TimeDuration maxResampleTime = std::min(touchDiff / int64_t(2), mMaxPredict);
     TimeStamp maxTimestamp = currentTouch.mTimeStamp + maxResampleTime;
     if (sampleTime > maxTimestamp) {
       sampleTime = maxTimestamp;
